Create Kings User Requirments
This commit includes the user requirements that have been set out for the application. They are all brief use cases and only include what is necessary for Kings game alone. No user requirements have been made for the possibility of other games been added.
These are created from a first look and the in reality the application may need more to sustain it.