Uploaded Kings UML.docx
I created the UML diagram in a word document as I did not have access Microsoft Viso. The tables are clear and show the variables and functions that I will have to create. As of now this is what I can think of for each class. More functions and variables will be added or taken away in the future. For now I think I have designed enough to make my next step coding. 
I will keep all versions of the UML and show newer versions as v2, v3 etc