In this commit minor changes were made to the 'Welcome Screen' and 'Game Details' forms. Only changes that were made were for names and location of text boxes. I had also made an error with the type of boxes I was using to display read only data. I was using text boxes when I should have been using Labels. New forms were created such as:           Kings           Add Player           Set Decks           Exclude Cards 'Kings' is the main form for playing upon. It contain 4 text boxes to display the card name, rule, cards left and decks left. It also has a picture box to show the picture of the curret card. A menu bar was also included to give users multiple options in the game. To go back to home page, add or remove players, add decks and restart the game.
The 'Add Player' form will show when the user tries to add a player to the game. It will prompt the user for the name and then give them the option to add or cancel.

The 'Set Decks' form will show when the user tries to add a deck. It will prompt the user to enter the amount of decks using the available options within the combo box. User then has option to set or cancel

The 'Exclude Cards' form will be shown when the user wishes to exclude a card. They will be given a combo box with all the cards in play and they will be allowed to choose which one they wish to exclude. Again they will have the option to exclude or to cancel when they have chosen the card.