Updates to forms and creation of new forms
In the menu of the 'Kings' form two items were added to the menu. 'Change Player Name' and 'Undo Exclusion'. They will provide links to new forms that were created for them. These forms will allow the user to slect the player name or the excluded card from a grid box and then slect the new name or choose to undo the exclusion of the card

Two Text boxes were added to the 'Kings' form to display the current player name and the amount of players that are in the game.