In this commit the classes 'Players' and 'Card' were created
The 'Card' class was pretty straight forward with its variables for cardName and cardRule. It consisted of a no arguement constructor and a multi arguement constructor as well as getters and setters. That is all this class consisted of.

The 'Players' class was more complex. Its variables were a string 'playerName' and a List<String> called 'playerList'. Again this class contained an no argurment and multi argurement constructor. It had functions as follows:
     addPlayer
     getPlayer
     nextPlayer
     removePlayer
     changePlayerName
     nameExists
'nameExits' and 'getPlayer' are new functions that were not included in the UML diagram. Once I began coding I saw that there was a need for these functions.
All functions work as they are called, either using loops tom transverse through the list or using functions inheriated from the list.
The main different between getPlayer and nextPlayer is that getCard will return the first vlaue of the list. While nextCard will read a value given by the user and then output the next value in the list. If it reaches the end of the list it will go back to the start of the list again.