Creation of 'Deck' Class
This is the most complex of the classes so far as it has the most control over the application. It has the most functions and processes the most amount of information.
My orignal idea was to create an enum to store the names of the cards and I coded it like this. It was when I tried to select a random value from the enum that I ran into trouble. I relised that it would be a lot of work to try and get a random value from the enum when I could just use a List<String> and get the random value very easily. I have not tested anything as of you but it should work in theroy. I created the list and added the names of all the cards of a deck. I then started the next card function, generted a random card and created a switch case to provide information for the card selected. As of now I have all the card names set in the switch but the rules of the cards are still wrong. These will have to be done in the next commit.