User Requirements for other possible games
This text file is the user requirements for other possible games that I made add to the application. I felt that the single game would not involve enough work and would not allow me to display my knowledge of coding so I may add these games. They will still use the card and deck classes but may force me to modify them with extra variables if I add the games. 
All the games have the similar idea but just differ on minor rules. Many of the forms already created can also be used for them. All that will be needed is an extra form to allow a user to choose their game and forms to be a UI for the extra games.