New Functions For 'Deck' class
In this commit the functions 'setdecks','addDeck' and 'removeDeck' are added to the 'Deck' class. There were not complicated classes and olny allowed the control of decks for the application. For seting the decks at the start, adding decks during the game or removing decks during the game.
'setDecks' will be given a float value and the number of decks will be set to this
'addDeck' will also be given a float value and this will be added to the number of decks
'removeDeck' will recieve a float value and then check if this value is less than the number of decks. If so it can be taken from the number of decks. Upon complitation this function will return a Boolean value based on the result

Changes were also made to the 'getNextCard' class. Each case contained a return statement, but I removed these and replaced them with a break. I also added a default case to set the card to null value. The single return statement was then shown at the bottom.