Further Updates To 'Deck' Class
In this commit I created the 'reduceCardsLeft','setExcludedCard' and 'resestGame'. The functions are rather complex and helped mre realise while I was coding them that I would need to rethink how I manage the deck due to the removal and number of decks. There would be serious problems if i removed a card and then tried to add it back. Especially when dealing with half decks. As of yet any of the functions created in this commit are not finished so I will explain their function better in the next commit. I feel they will change a lot as i am still trying to get my head around on what to do with this situation and dont have a clear idea yet.