Update to Functions of 'Deck' class
-joinDecks
This function was to be used internally by the Deck class to add the required amount of decks to the playable deck. This would let the same card be stored more than once. It can add both full decks and half decks

-setDecks
This function called on the joinDecks class to add the decks as it was setting the decks at the beginning of the game. The next part of the function ws to declare the number of decks in the numberOfDecks variable and insert the number of cards into the cardsLeft variable

-addDeck
This function also called on the joinDecks function to add the extra decks to the playable deck. It then added to the numberOfDecks and cardsLeft

-removeDeck
This function was to allow a user to remove decks if they wanted to shorten the game. This proved a difficult problem to solve as the  right cards had to be romved from the deck oir it would become inconsistant. The solution was similiar to the code in the joinDecks function. Use a loop  to iterate through the deck and remove the first instance of every card from the orignal deck. This would solve the problem of the list no longer being ordered. It could also be done for the first half of the deck as the user has the option to remove half a deck. It was done the same way except with a loop of 26.

-reduceCardsLeft
The modification to this class was to give it a input parameter of type int to specify the index that was to be deleted. This is the index of the card being used of the excluded card. I am not currently sure if this input parameter wll stay as it could cause problems in the excludedCard function.
The other tasks this function this is remove a card from the cardsLeft and adds to the cardsUsed and if the number of cardsUsed reaches 26 the numberOfDecks will be shortened by 0.5. cardsUsed will then be reset to 0.

Other modifications that were made was the creation of the OrignalDeckOfCards to allow an original deck to always be stored. This allowed mutiple OrignalDeccks to be added to the DeckOfCards

All other functions that are below the reduceCardsLeft are not yet finished and will be explained in a future commit