Updates to 'addDecks',excludeCard' functions of Deck Class
In this commit I edited the 'addDecks' function so when a new deck was added it would be checked for cards that had been excluded. It would then remove these cards. I used a loop that would keep removing the card until no occurences of it existed. This was more efficent than looping for the amount of decks in the game.

The 'excludeCard' function was was also changed to store a similar loop that would keep looping until all matching cards had been removed. Inside the loop it would remove the card, add the card to usedCards list, decrement cardsLeft and increment cardsUsed. When it had been confirmed that the card no longer existed within the deck the card would be added to the ExcludedCards list once.

All of these lists and functions will make reseting the game a lot easier as all necessary is now available. Hopefully the reset function will be completed in the next commit.