Creation of 'Game' class as well as updates to 'Deck' class
In the 'Deck' class I wrote the function for reseting the game. The main process is to add all the used cards back to the deck and reset the amount of cards used and cards left. It will then clear the UsedCards and ExcludedCards lists so they will be ready for the new game.

The 'Game' class is the final code class to be created. It acts as a link between the forms and the classes 'Deck' and 'Player' It creates its own functions based on the needs of the forms and calls upon the funcions of the 'Deck' and 'Player' class. The reason of this is to minimise the amount of code in the forms as much as possible. If possible I would just like functions to be called in the forms and nothing else.
The functions created are as follows:
     -NextCard
          >Simply gets the next card in the deck using the getNextCard function from the 'Deck' class
     -SetDecks
          >This will set the number of decks at the start of the game and using the setDecks function in the 'Deck' class it will set the number of cardsLeft.
     -AddDeck
          >This can be done throughout the game if the user would like to add another deck to the game. The functions in the 'Deck' class will take care of any excluded cards that need to be taken out etc
     -RemoveDeck
          >This is the opposite of the previous function and just allows a player to remove a deck or half a deck at any point in the game. This is only allowed if a sufficent number of decks are available
     -NumberOfDecks
          >This will return the number of decks that are currently in play
     -CardsLeft
          >This will show the number of cards that are yet to be drawn
     -ExcludeCard
          >This is a function a user can call on at any time in the game if they would like a certin card not to be included anymore
     -ResetGame
          >This will set the game back to the start based on the number of decks left. It will not return the orignal amount of decks but rather restore the remaining decks back to full
     -AddPlayer
          >Allows a user to add a player at any point in the game
     -SetPlayers
          >This is called at the start of the game to allow a bulk number of players to be entered at once.
     -SetExcludedCards
          >This is another function that is called at the start of the game to allow a bulk number of cards to be exluded from the game
     -FirstPlayer
          >This will return the first player of the player list
     -NextPlayer
          >Using the name of the previous player this will return the next player in the list
     -RemovePlayer
          >This will allow a user to remove a player at any point in the game
     -ChangePlayerName
          >This will allow a player to change the name of an existing player at any point in the game.

I have indentified two errors to be fixed in the next commit. Another form must be created for the entering of player names at the start of the game.  Currently the AddPlayer form will only work for adding players throughout the game.

Upon reseting the game, if the original amount of decks is not restored than adding the UsedCards list to the DeckOfCards list will cause problems. The solution might be adding the counts of both lists and seeing if it matches with the cards left. If not than they should be set equal