Updates to forms Welcome_Screen and Game_Details
This commit started with minor changes to the Deck class. This included creating a variable called OriginalDeckOfCards so when the game is reset it can be restored to its orignal state. I also created the updateCardsUsed function so anytime a card was used it would check if it was matching the number of decks left. As seen in the code if used cards reached 26 the number of decks would be reduced by 0.5.
I also created a function to undo exclusions. This simple just adds the excluded cards back to the deck and adjust the cards left and cards used accordingly.
In the game class a function was created to call upon the undo exclusion function

In the form Welcome_Screen I wrote code so that when the play button was hit it would go to the next form which was Game_Details.
In the Game_Details form I had much trouble with dragging data across forms. When I called upon the Add_Player form I would need the player name to be brought back to the Game_Details form. After a long time of thinking about it I simply iniated the Add_Player form in the Game_Details form and used it as a FormDialog. This allowed me to simply use it to get the name and close it. The name was then added to a datagridview and as many names as wanted could be added to the grid. The same was done for the number of decks and picking the excluded cards. All details were fed back to grid views. I created a single function to add to grid views with parameters for the value and name of the grid. I created a function for each dialog form that had to be called. All pretty much the same.

The current task at hand before I can go past the Game_details form is to remove a card from the comboBox for excluded cards once it has been excluded. I have researched it and the code I curerently have should work but nothing seems to happen. I will look at this more when I have time.