Created the startGame button functio
Made a function called CreateGame that will read into the DataGridViews to the Game class.
This also passed the game class to the Kings class so the data could be accessed.
Small changes were made to the SetPlayers and the SetExcludedCards because they were not reading the gridviews properly. They should be working properly now.