Updates To Kings class and correction of errors in other classes
The main code wrote in this commit was to the Kings class to control the output of data. The first step was to make sure the data was imported from the Game_Details class. This was done by giving the Kings class a Game arguement that would take the data generted and use it in the creation of the Kings form. The next step was to manage the data. So far I have created two classes that will be called on to fill in the card details and another class to fill in the player details. As there will be many points in the game where these details must be refreshed, these functions will help prevent data redundancy. As of now I have only called them on the loading of the form and when the next button is pressed.
A few errors were shown to me when trying to run the application. There were simple errors where a wrong data type was being converted or a value didnt have a .ToString after it. Other errors included the Player Amount being wrong as it was counting the empty row as well. And when the number of cards left reach 0 the application crashed as it was still trying to get cards when there was none left. This was solved by not allowing the next card function to retrieve a card if the cardsLeft was equal to 0.
Other errors included the name not showing on the kings page. This was cause Rows.Cell.Value was missing the value part and I was inserting blank values.
Also did a few resizes on the Text boxes of the kings form as I had made them too big.

In the next commit I hope to start and finished the menu of the kings form. This will probably produce a lot more errors when adding extra players and decks and excluding cards.