Adding Of Images As Items
This commit was just an updating of the names of each image so they can be more easily referenced. I then created a folder called Images and added the images as existing items. This allowed me to reference them and use them in the picture box.

I also called one image to be shown on the Kings for load. This worked at first but did not show the whole image due to size issues. Once I called on the SizeMode and set it to stretch the image the image displayed perfectly.

The next part of this process will be to add the image path to the details of each card so a different image will be called upon for each image.