Coding of Menu Buttons in Kings Form
The main part of this commit was to code the Menu Buttons in the Kings form. For the most part this was just calling on already created forms and functions.
This was the case for adding and removing players and adding decks. It was as simple as calling on the forms and functions.
For excluding a card and undoing the exclusion of a card it was a little more complicated. For this I had to write new functions in both Deck and Game class. I did this to return the Lists of excluded cards and the main deck of cards. These lists were then added to two combo boxes so the user could pick the card that they wich to exclude or un-exclude. I have not tested it yet so it only works in theroy.
I created a form for the RemovePlayer function.

Other changes I made were for the image of each card to show. As of the last commit I had the images ready to be used. In this commit I add the CardPath as a variable of the Card class. In the Deck class the path was changed for each different card making it easy to reference from the King class. It was as simple as calling the path and giving it to the image location and the image showed perfectly.