Fixing Of Errors And Creation Of Test Classes
The work that was done today involved some minor error checking. What I was mainly doing was running the project and making sure I could carry out each task without crashing the application or producing a wrong value. It was not a deep check but the main aspects of the app that would stand out to a user. This invloved the number of cards left not decrementing when a card was excluded, the application crashing when I tried to undo an exclusion, too many cards showing in the undo exclusion combo box, the first time I would add an extra deck it would not stay past the next go. It would reset to its previous value. These were all simple errors around the code that were very easy to fix and were just caused by a mistake in the logic. Once these had been complete the main coding of the project was done. The next stage was to carry out testing on all functions and make sure that they are working properly as it would be impossible to test this from just running the application and seeing how the answers compared.
I started by creating a test project and then a test class file. I created 4 classes insdie the file for each core class of the Kings Card Game. So far I have tested the getters and setters of the Card class. This is simple testing and it would be extremely rare for one of these tests to fail. This testing was just to help me understand tests again and start out with something basic so I would be ready for the more complex functions.