Updates To Tests of the Deck Class
The main coding of this commit took place in the testing class for the Deck class. This simply creating a test method for each method and then establishing a scenario where certain values would have to be the outcome or else the test would fail showing faults in the method. For most methods this worked perfectly and the methods worked just as they should have. This was the case except for two methods.
The first method to have an error was the RestartGame method. The error was that the number of decks was no being reset. Without the testing method this would have been very hard to notice. It was a simple fix of basically setting the number of decks equal to the original number of decks.
The second error that I cam across proved to be a lot more complcated to understand, but in the end it was another simple fix. It was the UndoExclusion method and for some reason the method was not going through its loop enough times. After some debugging it was easy to determine exactly where the error was. It was in the loop command itself. I original had a FOR loop but I changed it to a WHILE hoping it would solve the problem. However it did not and I was forced to keep looking. As far as I could see the code was sound and it should be looping normally. I was right in saying this but I did not take into account what was happening inside the loop. The loop would continue to loop until the variable i was equal to the UsedCards count. This was correct but I did not take into account that a card was removed the UsedCards list everytime it looped. Once I set the WHILE loop to <= the method was fixed. This was a complicted mistake to find but so easy to fix.

I finished testing the Deck class in this commit proving that it is good code and has no obvious errors. Of course more testing will be needed to test all scenarios.

Other updates to the project was the creation of getter methods in the Deck class. I created 3 methods just to return variables for the tests. GetOriginalDeck, GetUsedCards and GetCardsUsedCount. This were just simple getters to allow me to compare data.

I also made modifications to the code layout as it was very messy. I must do more modifications to the class as it can still be hard to understand what is happening based on the positions of the methods.
This type of modification will need to be done to all classes as they are all a but messy and need to be cleaned up so they will be more understandable for other coders.

In the next commit I will try to test the Player class or the Game class.