Testing of Player Class
This commit involved all the testing of the player class to assure that all information is being processed properly. There was only a few methods such as AddPlayer, NextPlayer, GetPlayerAmount, RemovePlayer, ChangePlayerName. I did not test GetPlayer as this is only a getter method. All these tests just involved creating a scenario and comparing the expected result to the scenario result. These tests did not produce any errors. It was during this testing I cam across the NameExists method and realised that it was never being used and I could not even be sure if it worked properly. Rather than testing a method that was of no use to me I simply deleted. In the future a method like this may be created again in a different way to help eliminate redundant code.
I also decided not to test the Game class as this would be usless testing. Each method of the Game class just calls methods from the Player and Deck class. Methods that have already been tested and are working perfectly. Due to this I decided to do no testing on the class and removed any code that had already been wrote.