Modifing Layout and Refactoring Code
The first thing I did for this commit was to modify the leyout of the classes so they would be easier to navigate through. It just involved grouping methods that had things in common and identifing them with comments. It was not a lot of work but I think it makes reading the code a lot easier for me and other outside users.

The next part of the commit involved code refactoring. Earlier in my commits I had stated that I wanted to have as little code as possible in my forms. I had kept with this apart from the Kings form. In this form the event methods for modifing the playername, be it adding, changin or removing had a lot of code in them. It was the same for excluding and un-excluding cards. They all contained similiar code on how to call a dailog box and then feed the results into a Game method. To fix this probelm I simply changed the Game methods to run the dialog boxs themselves. This removed most of the code from the event class and left the methods simply calling the Game methods.
I also thought about creating another method in the Game class to call the dialog box rather than having the same code repeated in 5 methods. But I did not think the amount of code being repeated warrented a seperate method to contain it. It was only one line per method that was similiar.