Refactoring Game_Details Class
In this commit I continued to refactor the code and eliminate code smells. A code smell that I found in the Game_Details form was methods that were executing code on behalf of the events. The events would call these methods when an event happened. It occured to me that these methods were out of place and I wanted as little code as possible in the form classes. To fix this I moved these methods to the Game class and then called them through the game class from the Game_Details class. The methods were AddToDataGrid, AddPlayerDialogBox, SetDecksDialogBox, ExcludeCardsDialogBox. A few changes had to be made after they were moved to carry over the data grids and make sure the mehtod could reference the correct forms and buttons. It was just adding extra parameters to the methods. This worked perfectly once complete.

I beleive this was also a benefit for the Game class as I was fearfull that it could be seen as a MiddleMan class. Even though the Game class does have great importance as a HUB for the Deck and Player class a lot of its methods were just calling methods from other classes.
Its methods like the ones moved in this commit that will make the Game class more important with individual methods of its own. After the refactoring in this commit and the previous I believe that the Game class has become a lot more important in function and is now more necessary for the application.