Error Fix on Player
During further testing of the application I found an error in the Player class. I never took into account what would happen if duplicate names were entered. If this was done the game could get stuck looping the names when the next card was called as it would always called the card after the first occurence of the name. This was a problem so it had to be fixed. After some  thought I tried to use a foreach to go through the list and if it found a matching name it would add a number to the current name. But it was not possible to sustain this for a 3rd duplicate or a 4rth duplicate etc
The next approach which took some time to figure involves counting the number of times a name exists and then adding the new name with a _2 or _3 added on to it to define it. The difficult part was adding this append on. A name with the append _2 could already exist so the append has to keep incrementing until it finds a unique value. I have now tested this solution and can say the problem has been fixed and duplicate values cannot exist.