Removal Of Variables
In the Game_Details I removed a call to the Add_Player form as I could not see it being used anywhere. The methods it might have been used in were now located in a different class

In the Game class I change the name of the list variable to SelectedCards. The naming convention was not great before and did not offer any explnation to what the list did.