Removel Of Methods
After my last commit where 3 of the core classes were made abstract and the Game class was able to inherit the methods from all the other classes I realized that I now had duplicate methods. Most of the methods I had created in the Game class were just calling on the methods from other classes and in some cases added an extra bit of code. Now that all methods have been inherited this was no longer needed. So I have moved the extra code that from the Game class to the Deck and Player class. I also changed the method names so calls from the Kings and Game_Details form will still be recognised. This made the methods of the Game class redundant so I simply deleted them. I am now left with 5 methods in the Game class that cannot be moved due to their function. This is a big change from the amount of methods I had before.
This was a code smell making the Game class a middle man, causing the code smell feature envy and also the code smell shotgun surgery in a way.

Minor changes were made to the Test class but these will be changed again and explained in the next commit