Testing Updates
For the Card class testing any calls on the Card class were replaced with call in the Game class cause all the methods can not be called through the game class.
For the Deck class I have been changing the tests as some of the tests no longer work since I changed the methods.The problem I am having right now is trying to test a method that has a dailog box. I have not figured out a way to automatically set the answer for the dialog box.