Tests Updates
Since the changes in my code the tests had to be updated to test the new formats of the methods. This was nothing major but only changes in input parameters and method names. A number of methods were also removed cause they could no longer be tested or they no longer existed.
A minor update was made in the SetDecks method. The code that set the number of cards was moved from the SetDecks method to the JoinDecks method. This was done because the JoinDecks test was failing.