Refactoring Deck Class
I have known for a while that the switch statement in the Deck class was violating two coding smells. Long Method and Switch Statement. I had convinced myself that I needed it and I could not change it. After further look I realised that I had to do something as this was bad code. I dedcied to use a dictionary to contain the rules and have the card name as the key. As the NextCard method was already generating the loop to find the card I replaced 400 lines of code with 3 lines. Simpyl by setting the card values everytime the next card method was run.  This was a smell I should have fixed earlier as it was one of the worst. I have since tested the project and everything is running smoothly.