Update Player Class
During testing I discovered that the problem of duplicates was not completly finsihed. While it could not happen when adding a player name it was still a possibility when changing a player name. My solution to this was to use the already existing method 'addPlayerToList' instead of the insert method. This would provide the same validation and not allow a duplicate name to occur and eliminate the risk of duplicate code in two methods.