Error Fixes and Updates
In my opinion I have now finished code refactoring. Code Analysis to be done with outside softwares, but the default softwares of Visual Studio cannot find anything.
This has lead me to do some error testing on my project. This is not using the test methods or class I created by manually messing with the application myself. I did this by creating a fake game with 5 players and 2 decks. I then played the game as if it was a real game. The aim was to take notice of anything that wasnt functioning right. I did not believe I would find much. #
The first error I found was when adding a deck during the game. While only one deck was added double the amount of cards would be added. This was because the cardsleft was being set twice. It was an easy fix to remove on of them.
The second error was in the JoinDecks method. Again the cardsleft was not being set right. I had set the new amount of cards to replace the old amount of cards when they should be added. It was a simple fix of changing a = to +=.
The third error I came across was an error with removing a player name. If the name of the current player was remove the applicaton would crash when the next player was called. This happened as the method could no longer reference the previous name. It was fixed with an IF statement to check if the name existed. If the name didnt exist it would simply return the first player and start from the start.

Other changes that were made to the code was setting the focus on certain buttons in the dialog forms. This was done to allow a player to simple hit the enter button to confirm rather than have to click it.
I also changed all the rule names to uppercase so they would stand out more when playing the game.