Validation
This commit involved making sure player inputs were validated and that they would not crash the application.  Using an IF statement I made sure that blank values could not be entered for adding a player name, changing a player name or removing a player name. I did the same for set decks and excluding cards to make sure that the player had to select a vlaue from the combo list and not leave it blank. I also changed the style of the combo boxes so a player was forced to select from the list and not enter their own data. This stops the player from entering bad inputs.